no point in non-blocking for fd notifications
It's not needed since the native thread will retry in the
unlikely case of EINTR/EAGAIN.  And writing one byte to a pipe
that's guaranteed by POSIX to be at least 512 bytes is highly
unlikely.

It's also bad because F_SETFL takes the big kernel lock under
Linux (and possibly other systems), and doing it unnecessarily
is a waste of system cycles.