do not release GVL when unlinking/opening
Since the message queue is not actually on a (slow) block
device, it's unlikely to block in a way where other tasks may be
scheduled by the kernel.  So avoid complicating things and
unnecessary task switching and assume mq_open/mq_unlink can
be executed as fast as the CPU/memory subsystems allows.