posix_mq 0.7.0 - cleanups + portability fixes
This release now works for systems other than Linux and FreeBSD
that support POSIX message queues.  This is usable for systems
without mq_timedsend() or mq_timedreceive() but with the
non-timed variants (timeouts wont work).

Thanks to Tadeusz Bochan for testing.

Ruby posix_mq is now dual-licensed under the GPLv2 (as well as the
preferred LGPLv3) for compatibility with existing GPLv2-only code.